fixed: resolve issue with email not being sent correctly to the user
