+++ backend/server.js
@@ -10,6 +10,7 @@ const path = require("path");
 
 const user = require("./models/user");
 const db = require("./config/databaseConfig");
+const FCEOMember = require("./models/fceomember");
 
 require("./associations/association").eventAssociations();
 require("./associations/association").competitionAssociations();
@@ -30,7 +31,7 @@ db.authenticate()
   });
 
 // Database Table synchronizing
-// db.sync({ alter: true })
+// FCEOMember.sync()
 //   .then(() => {
 //     console.log("Update Database");
 //   })
